style: Move AIR tests to `air.rs`
